pyproject.toml
[project]
name = "media-discord-rpc"
version = "0.1.0"
description = "Display your listening activity from subsonic-compatible servers & watching activity from jellyfin."
readme = "README.md"
authors = [
    { name = "Karn Wong", email = "karn@karnwong.me" }
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "jellyfin-apiclient-python>=1.11.0",
    "pypresence",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project.scripts]
media-discord-rpc = "media_discord_rpc.cli:cli"

[tool.uv.sources]
pypresence = { url = "https://github.com/qwertyquerty/pypresence/archive/master.zip" }
